chore: move typing out of _internal
