+++ src/boost_histogram/_internal/view.py
@@ -5,7 +5,7 @@ from typing import Any, Callable, ClassVar, Literal, Mapping, MutableMapping
 import numpy as np
 
 from ..accumulators import Mean, WeightedMean, WeightedSum
-from .typing import ArrayLike, StrIndex, Ufunc
+from ..typing import ArrayLike, StrIndex, Ufunc
 
 UFMethod = Literal["__call__", "reduce", "reduceat", "accumulate", "outer", "at"]
