+++ src/boost_histogram/typing.py
